Extract out common base session wizard form
